Рефакторинг страницы коктейля
